requirements.txt
pyspellchecker
faker
